+++ packages/tools/utils/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 - updated docs
 
+### Any -> (Array -> Object)
+
+- added `makeKeyed`
+
 ### Array -> Array
 
 - added `getFirstAndLast`